Implement orphaned files removal feature (#75)
The problem of orphaned files arose when a table was created and data was loaded
within an active transaction. If a critical failure occurred before the
transaction could be committed or aborted (e.g., a sudden system power loss or
an unexpected shutdown of a database node), the system underwent a crash
recovery process. While the logical table was correctly marked as absent or
rolled back, the physical data files associated with that uncommitted table
could persist on the underlying file system. Over time, these orphaned files
could accumulate, cluttering storage space and leading to unnecessary resource
consumption.

Prevously, addressing this required manual intervention for cleanup. The
feature, introduced in this patch, allows to automatically remove such files
during crash recovery.

The idea behind the feature is as follows:
1. Each backend process stores information about files created for new relations
('relfilenode's), along with their current transaction ID, in its own internal
list. Once the transaction is committed or aborted, all relfilenode entries for
that transaction are removed from the list.
2. The checkpointer process aggregates information from all such lists across
all backends and stores the accumulated data in a new, dedicated WAL record
XLOG_PENDING_DELETE.
3. During crash recovery, the Startup process, which replays the WAL, obtains
the information about created relfilenodes from XLOG_PENDING_DELETE and
XLOG_SMGR_CREATE records and stores this info in a hash map. On the replay
of transaction commit or abort, the respective relfilenodes are removed from the
hash map. Relfilenodes for prepared transactions are removed from the hash map
as well. Finally, the crash recovery process deletes the physical files for the
remaining relfilenodes in the hash map for all transactions that are marked as
still in progress.

(cherry picked from commit 3ef385f4464e6486f71fcc52dceb3cde66af462e)
(cherry picked from commit 780483134615792d628581bf8626c418e69ca246)
(cherry picked from commit ceea92ee5d6576bdcb5c38a63507acd9dc44f3ed)

Major changes compared to the original commit:
1. XLOG_SMGR_CREATE_PDL WAL record, introduced in the cherry-picked commit, is
no more needed, as the existing XLOG_SMGR_CREATE WAL record now contains all
required info about relation storage type (via 'SMgrImpl impl' field). So,
XLOG_SMGR_CREATE_PDL WAL record is removed from the commit, and related changes
in
 - 'extractPageInfo()';
 - 'smgr_desc()';
 - 'log_smgrcreate()'
are also not needed. As parameters of 'log_smgrcreate()' didn't change, all
delta related to the calls of this function is also no more relevant. Logic in
'smgr_redo()' for XLOG_SMGR_CREATE is updated to include 'PdlRedoAdd()' as it
was done for XLOG_SMGR_CREATE_PDL. Structure 'RelFileNodePendingDelete' in 7x
has 'int smgr_which;' instead of 'char relstorage;' - so the related code is
updated as well.

2. Function 'LWLockAssign()' is absent in 7x. So
 - use 'LWLockInitialize()' to init locks in PdlShmemInit();
 - add new LWTRANCHE_PENDING_DELETES internal tranche id to use with
'LWLockInitialize()';
 - in struct 'PendingDeletesList' change 'LWLock *lock' to 'LWLock lock'.

3. 'PdlXLogInsert()' is changed because of the updated approach for WAL record
insertion - now one needs to call 'XLogBeginInsert()', 'XLogRegisterData()' and
finally 'XLogInsert()' instead of a single 'XLogInsert()'.

4. Handling of prepared transactions is significantly reworked in 7x. Therefore,
'RemovePendingDeletesForPreparedTransactions()' internal logic is also reworked.
As now the information of a prepared transaction is moved from WAL into a
pg_twophase file during the checkpoint, the situation with recycled WAL files
(when some prepared transaction is not yet committed/aborted, but the respective
WAL segment file is already recycled) should not be relevant. Because of it, now
'RemovePendingDeletesForPreparedTransactions()' doesn't need to provide a return
value, and we call 'PdlRedoDropFiles()' in 'StartupXLOG()' right after it
unconditionally.

5. 'xlog_redo()' in 7x has 'XLogReaderState *record' parameter instead of
'XLogRecord *record', so the parameter of 'PdlRedoXLogRecord()' (called from
'xlog_redo()') is also changed to 'XLogReaderState *record'. Unit tests are
updated accordingly.

Other changes (not so significant):
6. Function 'xact_redo_commit_internal()' is absent, so call to
'PdlRedoRemoveTree()' is moved to 'xact_redo_commit()'.
7. Function 'xact_redo_distributed_commit()' now calls 'xact_redo_commit()' for
the major part of the work, so there is no need to call 'PdlRedoRemoveTree()'
in 'xact_redo_distributed_commit()'.
8. The value of XLOG_PENDING_DELETE is changed from 0xD0 to 0xF0 (as 0xD0 is now
assigned to XLOG_NEXTGXID).
9. In 'xact_redo_commit()' and 'xact_redo_abort()' params of
'PdlRedoRemoveTree()' are changed to use 'parsed->subxacts' and
'parsed->nsubxacts'.
10. Removed check for 'dynamic_shared_memory_type' in 'is_tracking_enabled()'
in storage_pending_deletes.c, as DSA is always on in 7x.
11. In 'dsa_create_in_place()' param 'tranche_name' is now absent. The
respective change is done in 'PdlShmemInit()'.
12. Add preprocessor checks for FRONTEND into 'storage_pending_deletes.h', as it
is also included from the frontend's code (xlogdesc.c), while 'atomic.h'
included from 'dsa.h' included from 'storage_pending_deletes.h' has
a compilation #error if it is included in the frontend code.
13. ABI-check ignore files are not introduced, as they are already in the
branch.
14. Some changes we made in mock.mk and fmgrtab_mock.c to allow linking of the
new unit tests.
15. In the unit tests: 'test_shmem_type' is removed as DSM_IMPL_NONE is absent
in 7x.
16. In the tests: remove include 'helpers/server_helpers.sql' as it is absent
in 7x.
17. In the tests: in 'force_mirrors_to_catch_up_with_exception()'
'pg_switch_xlog()' is changed to 'pg_switch_wal()'.
18. In the tests: 'createTables()' function is updated in order to properly get
the names for index files of pg_ao_visimap and pg_aoblkdir. Previously this
information was obtained from the pg_appendonly table. In 7x pg_appendonly
doesn't have this info, so we get it from pg_depend.
19. In the tests: reduced the expected file count for segment files of the AO
tables. In 6x the new segment files were created for the first insertion of data
after the table creation, but in 7x the same segment files are used as long as
the insertion is done in the same transaction as table creation.
20. In the tests: add additional check of subtransactions into the isolation2
test with the prepared transaction.

Co-authored-by: Andrey Sokolov <a.sokolov@arenadata.io>
Co-authored-by: Roman Eskin <r.eskin@arenadata.io>